refactor: reuse verifier transcript appenders
