+++ crates/jolt-verifier/src/stages/mod.rs
@@ -28,6 +28,52 @@ pub mod stage8;
 #[doc(hidden)]
 pub mod zk;
 
+#[cfg(test)]
+pub(crate) mod test_support {
+    use ark_serialize::CanonicalSerialize;
+    use jolt_field::{Fr, FromPrimitiveInt};
+    use jolt_transcript::{FsAbsorb, FsChallenge};
+
+    #[derive(Clone, Default)]
+    pub(crate) struct RecordingTranscript {
+        pub(crate) chunks: Vec<Vec<u8>>,
+    }
+
+    impl FsAbsorb for RecordingTranscript {
+        fn absorb<T: CanonicalSerialize>(&mut self, value: &T) {
+            let mut bytes = Vec::new();
+            if value.serialize_compressed(&mut bytes).is_err() {
+                unreachable!("canonical serialization into Vec succeeds");
+            }
+            self.chunks.push(bytes);
+        }
+
+        fn absorb_slice<T: CanonicalSerialize>(&mut self, values: &[T]) {
+            let mut bytes = Vec::new();
+            for value in values {
+                if value.serialize_compressed(&mut bytes).is_err() {
+                    unreachable!("canonical serialization into Vec succeeds");
+                }
+            }
+            self.chunks.push(bytes);
+        }
+
+        fn absorb_bytes(&mut self, bytes: &[u8]) {
+            self.chunks.push(bytes.to_vec());
+        }
+    }
+
+    impl FsChallenge<Fr> for RecordingTranscript {
+        fn challenge(&mut self) -> Fr {
+            Fr::from_u64(0)
+        }
+
+        fn challenge_scalar(&mut self) -> Fr {
+            Fr::from_u64(0)
+        }
+    }
+}
+
 /// Build the one-hot [`JoltFormulaDimensions`] from the proof's one-hot config and
 /// the verifier-trusted geometry (trace length, lookup operand width, bytecode
 /// length, RAM size), mapping the layout error to `stage`. Shared by the stages
